hello/requirements_release.txt
grpcio==1.28.1
grpcio-tools==1.28.1
pkg-resources==0.0.0
protobuf==3.11.3
six==1.14.0
